started work on hmd- display support
this branch adds support for the hmd_server tool that is launched and
managed by arcan. It adds support for aggregating numerous devices into
a 'sensor fusion' that provides information on a number of limbs that
can be mapped to an avatar, along with some minor control commands
and metadata for activating display and setting up distortion.